requirements.txt
fastapi
uvicorn
sqlalchemy
tornado
jinja2
python-dotenv
psycopg2-binary
pika